Moving the scheduling plugins under epp/framework/plugins
