+++ conformance/tests/gateway_following_epp_routing_dp.go
@@ -35,7 +35,7 @@ import (
 
 	"sigs.k8s.io/gateway-api-inference-extension/conformance/resources"
 	k8sutils "sigs.k8s.io/gateway-api-inference-extension/conformance/utils/kubernetes"
-	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/scheduling/framework/plugins/test"
+	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/framework/plugins/scheduling/test"
 )
 
 // dpPorts are the data parallel ports exposed by the backend deployment. Update if the ports
