+++ pkg/epp/framework/plugins/scheduling/scorer/predictedlatency/running_request_queue.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package predicted_latency
+package predictedlatency
 
 import (
 	"container/heap"
