+++ pkg/epp/framework/plugins/scheduling/scorer/predictedlatency/scorer_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package predicted_latency
+package predictedlatency
 
 import (
 	"context"
